fix(backend): Ensure all existing records have a unique UUID
This is somewhat complicated as we already have existing records, that
need to be updated to add a UUID to them. Having been given the process
to follow, I have now updated the existing migration to add the UUID field
and added some extra ones to add and populate the field to existing log
entries.

IssueID #40