feat(backend): added swagger_fake decorator (#44)
* feat(backend): added swagger_fake decorator

Added a decorator that can be applied to views to prevent swagger errors
during schema generation.  Certain views rely on logged-in users and/or
actual kwargs being passed to the views; those aren't provided during
swagger schema generation and so errors will ocurr.  There is already
code in several apps to stop further processing in those views.  But
doing it via a decorator will remove loads of code duplication.