fix(backend): Ensure logging models are unique
I made a mistake previously, the migration was made with the log
record set to be `unique`, but it then got removed and the migration
not updated. I have now added unique back to the model.

I also noticed that the log tag isn't unique, this is less of an issue
as checks are done first, but I think it is still good practice to
make the `name` field unique.

IssueID #45