fix(backend): fixed IncludeExcludeAdminListFilter
The updated query_strings used by IncludeExcludeAdminListFilter were not
being set correctly, leading to inconsistent behaviour when trying to
filter in the admin.  This was b/c the code to update the query_strings
was being called from w/in a generator so all previous updates were
being preserved from the last time a lookup_choice was generated.
Operating on a copy of the selected filter items fixed the problem.